// client side/src/interfaces/Card.ts
export default interface Card {

    _id?: string;
    owner: string,
    title: string;
    sutitle: string;
    description: string;
    phone: string;
    email: string;
    web?: string;
    image: string;
    state?: string;
    country: string;
    city: string
    street: string
    Hosenumber?: number;
    zip?: number;

}